feature: add servicelayer test utils
